move entry point, adapt tests
