--- lib/copperegg.rb
@@ -1,5 +1,0 @@
-module Copperegg
-  require 'copperegg/alerts/schedule'
-  require 'copperegg/alerts/client'
-  require 'copperegg/alerts/version'
-end
