+++ lib/copperegg/alerts.rb
@@ -1,0 +1,8 @@
+require 'copperegg/alerts/client'
+require 'copperegg/alerts/schedule'
+require 'copperegg/alerts/version'
+
+module Copperegg
+  module Alerts
+  end
+end
